+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.4.23
+
+* `Versions`:
+    * `Ktor`: `1.5.0` -> `1.5.1`
+* `Serialization`
+    * `Base64`
+        * New serializer `Base64BytesToFromStringSerializer` has been added
+
 ## 0.4.22
 
 * `Versions`: